Trial changes

+++ finetuning/models.py
@@ -608,19 +608,21 @@ class PairwiseWithOriginalDataJointTraining(L.LightningModule):
                 X1, X2, Y = (
                     dl_batch["seq1"],
                     dl_batch["seq2"],
-                    dl_batch["z_diff"].half(),
+                    dl_batch["z_diff"],
+                    # dl_batch["z_diff"].half(),
                 )
                 X = torch.cat([X1, X2], dim=0)
                 X = rearrange(X, "S H L -> (S H) L")
                 X = seq_indices_to_one_hot(X)  # (S * H, L, 4)
-                X = X.half()
+                # X = X.half()
                 Y_hat = self(X)
                 Y_hat = Y_hat[: X1.shape[0]] - Y_hat[X1.shape[0] :]
                 mse_loss = self.mse_loss(Y_hat, Y)
                 self.log("train/pairwise_mse_loss", mse_loss)
                 loss += mse_loss
             elif i == 1:  # this is the original human training data
-                X, Y = dl_batch["seq"].half(), dl_batch["y"].half()
+                X, Y = dl_batch["seq"], dl_batch["y"]
+                # X, Y = dl_batch["seq"].half(), dl_batch["y"].half()
                 Y_hat = self(
                     X, return_base_predictions=True, base_predictions_head="human"
                 )
@@ -628,7 +630,8 @@ class PairwiseWithOriginalDataJointTraining(L.LightningModule):
                 self.log("train/human_poisson_loss", poisson_loss)
                 loss += poisson_loss
             elif i == 2:  # this is the original mouse training data
-                X, Y = dl_batch["seq"].half(), dl_batch["y"].half()
+                X, Y = dl_batch["seq"], dl_batch["y"]
+                # X, Y = dl_batch["seq"].half(), dl_batch["y"].half()
                 Y_hat = self(
                     X, return_base_predictions=True, base_predictions_head="mouse"
                 )
@@ -648,19 +651,21 @@ class PairwiseWithOriginalDataJointTraining(L.LightningModule):
             X1, X2, Y = (
                 batch["seq1"],
                 batch["seq2"],
-                batch["z_diff"].half(),
+                batch["z_diff"],
+                # batch["z_diff"].half(),
             )
             X = torch.cat([X1, X2], dim=0)
             X = rearrange(X, "S H L -> (S H) L")
             X = seq_indices_to_one_hot(X)  # (S * H, L, 4)
-            X = X.half()
+            # X = X.half()
             Y_hat = self(X)
             Y_hat = Y_hat[: X1.shape[0]] - Y_hat[X1.shape[0] :]
             mse_loss = self.mse_loss(Y_hat, Y)
             self.log("val/pairwise_mse_loss", mse_loss)
 
         elif dataloader_idx == 1:  # this is the original human training data
-            X, Y = batch["seq"].half(), batch["y"].half()
+            X, Y = batch["seq"], batch["y"]
+            # X, Y = batch["seq"].half(), batch["y"].half()
             Y_hat = self(X, return_base_predictions=True, base_predictions_head="human")
             poisson_loss = self.poisson_loss(Y_hat, Y)
             self.log("val/human_poisson_loss", poisson_loss)
@@ -671,7 +676,8 @@ class PairwiseWithOriginalDataJointTraining(L.LightningModule):
             self.human_metrics["mse"].update(Y_hat, Y)
 
         elif dataloader_idx == 2:  # this is the original mouse training data
-            X, Y = batch["seq"].half(), batch["y"].half()
+            X, Y = batch["seq"], batch["y"]
+            # X, Y = batch["seq"].half(), batch["y"].half()
             Y_hat = self(X, return_base_predictions=True, base_predictions_head="mouse")
             poisson_loss = self.poisson_loss(Y_hat, Y)
             self.log("val/mouse_poisson_loss", poisson_loss)
